backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
pandas>=2.2.0
requests>=2.32.0
yfinance>=0.2.54
